# apps/api/src/modules/newsletter/controllers/newsletter.controller.ts
import { Public } from '@api/core/auth/decorators/public.decorator';
import { CurrentUser } from '@api/core/auth/decorators/current-user.decorator';
import type { AuthenticatedUser } from '@api/core/auth/interfaces/authenticated-user.interface';
import { ErrorResponseDto } from '@api/core/http/dto/error-response.dto';
import { RATE_LIMITS } from '@api/core/http/security/http-security.constants';
import { ConfirmSubscriptionDto } from '@api/modules/newsletter/dto/request/confirm-subscription.dto';
import { SubscribeNewsletterDto } from '@api/modules/newsletter/dto/request/subscribe-newsletter.dto';
import { UnsubscribeDto } from '@api/modules/newsletter/dto/request/unsubscribe.dto';
import {
  SubscriptionAcceptedResponseDto,
  SubscriptionConfirmedResponseDto,
} from '@api/modules/newsletter/dto/response/subscription-response.dto';
import { NewsletterService } from '@api/modules/newsletter/services/newsletter.service';
import { Body, Controller, HttpCode, HttpStatus, Post } from '@nestjs/common';
import {
  ApiAcceptedResponse,
  ApiBadRequestResponse,
  ApiGoneResponse,
  ApiNoContentResponse,
  ApiOkResponse,
  ApiOperation,
  ApiTags,
  ApiTooManyRequestsResponse,
  ApiUnprocessableEntityResponse,
} from '@nestjs/swagger';
import { Throttle } from '@nestjs/throttler';

@Throttle({ default: RATE_LIMITS.newsletter })
@ApiTags('Newsletter')
@ApiTooManyRequestsResponse({
  description: 'Limite de solicitações da newsletter excedido.',
  type: ErrorResponseDto,
})
@Controller('newsletter/subscriptions')
export class NewsletterController {
  constructor(private readonly newsletterService: NewsletterService) {}

  @Post()
  @Public()
  @HttpCode(HttpStatus.ACCEPTED)
  @ApiOperation({ summary: 'Solicita inscrição com double opt-in' })
  @ApiAcceptedResponse({ type: SubscriptionAcceptedResponseDto })
  @ApiUnprocessableEntityResponse({ description: 'Dados inválidos.', type: ErrorResponseDto })
  subscribe(@Body() dto: SubscribeNewsletterDto): Promise<SubscriptionAcceptedResponseDto> {
    return this.newsletterService.subscribe(dto);
  }

  @Post('confirm')
  @Public()
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Confirma uma inscrição pendente' })
  @ApiOkResponse({ type: SubscriptionConfirmedResponseDto })
  @ApiBadRequestResponse({ description: 'Token inválido.', type: ErrorResponseDto })
  @ApiGoneResponse({ description: 'Token expirado.', type: ErrorResponseDto })
  confirm(@Body() dto: ConfirmSubscriptionDto): Promise<SubscriptionConfirmedResponseDto> {
    return this.newsletterService.confirm(dto);
  }

  @Post('unsubscribe')
  @Public()
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ summary: 'Cancela uma inscrição de forma idempotente' })
  @ApiNoContentResponse()
  unsubscribe(@Body() dto: UnsubscribeDto): Promise<void> {
    return this.newsletterService.unsubscribe(dto);
  }

  @Post('account')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ summary: 'Inclui uma conta confirmada na newsletter de forma idempotente' })
  @ApiNoContentResponse()
  subscribeAccount(@CurrentUser() user: AuthenticatedUser): Promise<void> {
    return this.newsletterService.subscribeConfirmedAccount(user.email);
  }
}