# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/newsletter/subscriptions": {
      "post": {
        "operationId": "newsletter_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeNewsletterDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionAcceptedResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": "VALIDATION_ERROR",
                  "details": null,
                  "message": "Dados inválidos.",
                  "path": "/api/v1/newsletter/subscriptions",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 400,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Dados ou parâmetros inválidos.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Limite de solicitações da newsletter excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/newsletter/subscriptions",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Solicita inscrição com double opt-in",
        "tags": ["Newsletter"]
      }
    },
    "/api/v1/newsletter/subscriptions/confirm": {
      "post": {
        "operationId": "newsletter_confirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionConfirmedResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "410": {
            "description": "Token expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Limite de solicitações da newsletter excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/newsletter/subscriptions/confirm",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Confirma uma inscrição pendente",
        "tags": ["Newsletter"]
      }
    },
    "/api/v1/newsletter/subscriptions/unsubscribe": {
      "post": {
        "operationId": "newsletter_unsubscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribeDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": "VALIDATION_ERROR",
                  "details": null,
                  "message": "Dados inválidos.",
                  "path": "/api/v1/newsletter/subscriptions/unsubscribe",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 400,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Dados ou parâmetros inválidos.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Limite de solicitações da newsletter excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/newsletter/subscriptions/unsubscribe",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Cancela uma inscrição de forma idempotente",
        "tags": ["Newsletter"]
      }
    },
    "/api/v1/newsletter/subscriptions/account": {
      "post": {
        "operationId": "newsletter_subscribeAccount",
        "parameters": [],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Limite de solicitações da newsletter excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/newsletter/subscriptions/account",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Inclui uma conta confirmada na newsletter de forma idempotente",
        "tags": ["Newsletter"]
      }
    },
...
  },
  "info": {
    "title": "Vavito Archives API",
    "description": "Contrato HTTP da API do Vavito Archives. Corpos JSON e URL-encoded aceitam no máximo 1 MiB; uploads multipart possuem limites próprios por rota.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token do Supabase Auth.",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "email"
          },
          "reason": {
            "type": "string",
            "example": "INVALID_EMAIL"
          }
        },
        "required": ["field", "reason"]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Dados inválidos."
          },
          "details": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-12T20:15:00.000Z",
            "format": "date-time"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/newsletter/subscriptions"
          },
          "requestId": {
            "type": "string",
            "example": "019c2d62-6e90-7000-8000-000000000000"
          }
        },
        "required": ["statusCode", "code", "message", "details", "timestamp", "path", "requestId"]
      },
...
      "SubscriberConsentSource": {
        "type": "string",
        "enum": ["ACCOUNT", "ARTICLE", "FOOTER", "HOME"]
      },
      "SubscribeNewsletterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "leitor@example.com",
            "maxLength": 320
          },
          "consent": {
            "type": "boolean",
            "enum": [true],
            "example": true
          },
          "source": {
            "example": "ARTICLE",
            "allOf": [
              {
                "$ref": "#/components/schemas/SubscriberConsentSource"
              }
            ]
          }
        },
        "required": ["email", "consent", "source"]
      },
      "SubscriptionAcceptedResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Se o endereço puder receber a newsletter, enviaremos as próximas instruções."
          }
        },
        "required": ["message"]
      },
      "ConfirmSubscriptionDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token opaco recebido no email de confirmação.",
            "example": "HfByP6b1hQ9lBf8Nw5vVJdWxe_vf1EpfkNGYw1iHt7Q"
          }
        },
        "required": ["token"]
      },
      "SubscriptionConfirmedResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Inscrição confirmada com sucesso."
          }
        },
        "required": ["message"]
      },
      "UnsubscribeDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token opaco recebido nos emails da newsletter.",
            "example": "HfByP6b1hQ9lBf8Nw5vVJdWxe_vf1EpfkNGYw1iHt7Q"
          }
        },
        "required": ["token"]
      },
...
    }
  }
}